requirements.txt
Flask>=3.0.0
pdfplumber>=0.11.0
openpyxl>=3.1.0
Jinja2>=3.1.0
Werkzeug>=3.0.0
Pillow>=10.0.0
gunicorn>=21.0.0
